refactor(types): move dep types to src/dep-types
