+++ packages/vite/package.json
@@ -116,6 +116,7 @@
     "strip-literal": "^0.4.0",
     "tsconfck": "^2.0.1",
     "tslib": "^2.4.0",
+    "dep-types": "link:./src/dep-types",
     "types": "link:./types",
     "ufo": "^0.8.5",
     "ws": "^8.8.1"
