+++ packages/vite/src/dep-types/alias.d.ts
