+++ packages/vite/src/dep-types/anymatch.d.ts
