+++ packages/vite/src/dep-types/chokidar.d.ts
