+++ packages/vite/src/dep-types/commonjs.d.ts
