+++ packages/vite/src/dep-types/dynamicImportVars.d.ts
