+++ packages/vite/src/dep-types/http-proxy.d.ts
