+++ packages/vite/src/dep-types/shims.d.ts
