+++ packages/vite/src/dep-types/terser.d.ts
