+++ packages/vite/src/dep-types/ws.d.ts
