+++ packages/vite/src/node/http.ts
@@ -5,7 +5,7 @@ import type {
   OutgoingHttpHeaders as HttpServerHeaders
 } from 'node:http'
 import type { ServerOptions as HttpsServerOptions } from 'node:https'
-import type { Connect } from 'types/connect'
+import type { Connect } from 'dep-types/connect'
 import colors from 'picocolors'
 import { isObject } from './utils'
 import type { ProxyOptions } from './server/middlewares/proxy'
