+++ packages/vite/src/node/plugins/terser.ts
@@ -1,5 +1,5 @@
 import { Worker } from 'okie'
-import type { Terser } from 'types/terser'
+import type { Terser } from 'dep-types/terser'
 import type { Plugin } from '../plugin'
 import type { ResolvedConfig } from '..'
 import { requireResolveFromRootWithFallback } from '../utils'
