+++ packages/vite/src/node/server/middlewares/error.ts
@@ -1,6 +1,6 @@
 import colors from 'picocolors'
 import type { RollupError } from 'rollup'
-import type { Connect } from 'types/connect'
+import type { Connect } from 'dep-types/connect'
 import strip from 'strip-ansi'
 import type { ErrorPayload } from 'types/hmrPayload'
 import { pad } from '../../utils'
