+++ packages/vite/src/node/server/middlewares/proxy.ts
@@ -1,8 +1,8 @@
 import type * as http from 'node:http'
 import type * as net from 'node:net'
 import httpProxy from 'http-proxy'
-import type { Connect } from 'types/connect'
-import type { HttpProxy } from 'types/http-proxy'
+import type { Connect } from 'dep-types/connect'
+import type { HttpProxy } from 'dep-types/http-proxy'
 import colors from 'picocolors'
 import { HMR_HEADER } from '../ws'
 import { createDebugger, isObject } from '../../utils'
