+++ packages/vite/src/node/server/middlewares/spaFallback.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs'
 import path from 'node:path'
 import history from 'connect-history-api-fallback'
-import type { Connect } from 'types/connect'
+import type { Connect } from 'dep-types/connect'
 import { createDebugger } from '../../utils'
 
 export function spaFallbackMiddleware(
