+++ packages/vite/src/node/server/ws.ts
@@ -6,7 +6,7 @@ import type { Socket } from 'node:net'
 import colors from 'picocolors'
 import type { ServerOptions, WebSocket as WebSocketRaw } from 'ws'
 import { WebSocketServer as WebSocketServerRaw } from 'ws'
-import type { WebSocket as WebSocketTypes } from 'types/ws'
+import type { WebSocket as WebSocketTypes } from 'dep-types/ws'
 import type { CustomPayload, ErrorPayload, HMRPayload } from 'types/hmrPayload'
 import type { InferCustomEventPayload } from 'types/customEvent'
 import type { ResolvedConfig } from '..'
