+++ packages/vite/src/node/tsconfig.json
@@ -1,6 +1,12 @@
 {
   "extends": "../../tsconfig.base.json",
-  "include": ["./", "../../types"],
+  "include": [
+    "./",
+    "../dep-types",
+    "../../types/customEvent",
+    "../../types/hmrPayload",
+    "../../types/importGlob"
+  ],
   "exclude": ["**/__tests__"],
   "compilerOptions": {
     "lib": ["ESNext", "DOM"]
